+++ settings.py
@@ -52,8 +52,7 @@ SPF_ATTRIBUTE = "hop"
 UNI_STATE_CHANGE_DELAY = 0.1
 
 # Default EVC values, they will be applied to every EVC which key is missing
-# E.g. "secondary_constraints": {"mandatory_metrics": {"ownership": "red"}}
 EVC_DEFAULT = {
-    "primary_constraints": {},
-    "secondary_constraints": {}
+    "primary_constraints": {"undesired_links": 123},
+    "secondary_constraints": {"mandatory_metrics": {"ownership": "red"}}
 }